- abilities should be done now

+++ common/src/main/java/smartin/miapi/modules/properties/util/DoubleOperationResolvable.java
@@ -58,7 +58,7 @@ public class DoubleOperationResolvable {
         this.operations = operations;
     }
 
-    public double getValue() {
+    public int getValue() {
         return evaluate(baseValue, fallback);
     }
